requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
flask>=2.0.0
pillow>=9.0.0
numpy>=1.24.0
streamlit>=1.22.0
huggingface_hub>=0.16.0
opencv-python>=4.7.0
scipy>=1.10.0
tqdm>=4.65.0
